refactor: NetworkManager 타입 내부 메서드 수정
