+++ BoxOffice/Controllers/ViewController.swift
@@ -17,15 +17,14 @@ class ViewController: UIViewController {
     }
     
     func networkTest() {
-        networkManager.fetchMovie(complitionHandler: { boxOffice in
-            boxOffice?.boxOfficeResult.dailyBoxOfficeList.forEach { dump($0) }
-        })
-        networkManager.fetchMovie { boxOffice in
-            guard let a = boxOffice?.boxOfficeResult.dailyBoxOfficeList[0].movieCode else { return }
-            self.networkManager.fetchMovieInformation(movieCode: a) { movie in
-                dump(movie?.movieDetailResult.movieDetail.self)
+        networkManager.fetchMovie(url: MovieURL.kobisURLString, type: BoxOffice.self, complitionHandler: { boxOffice in
+            boxOffice?.boxOfficeResult.dailyBoxOfficeList.forEach { /*dump($0)*/ $0 }
+            guard let code = boxOffice?.boxOfficeResult.dailyBoxOfficeList[0].movieCode else { return }
+            self.networkManager.fetchMovie(url: MovieURL.detailURLString(movieCode: code), type: Movie.self) { movie in
+                dump(movie?.infomationResult.movieInfomation)
             }
-        }
+        })
     }
+    
 }
