+++ BoxOffice/Network/NetworkManager.swift
@@ -8,43 +8,14 @@
 import Foundation
 
 struct NetworkManager {
-    enum DateFormat {
-        static var yyyyMMdd = "yyyyMMdd"
-    }
-    
-    func fetchMovie(complitionHandler: @escaping (BoxOffice?) -> Void) {
-        let urlString = "\(MovieURL.kobisURL)key=\(Key.key)&targetDt=\(fetchTodayDate())"
-        
-        executeRequest(with: urlString) { data in
-            guard let data = data else { return }
-            let safeData = DataDecoder.parseJson(type: BoxOffice.self, data: data)
-            complitionHandler(safeData)
+    func fetchMovie<T: Decodable>(url: String, type: T.Type, complitionHandler: @escaping (T?) -> Void) {
+        executeRequest(with: url, type: type) { safeData in
+            guard let data = safeData else { return }
+            complitionHandler(data)
         }
     }
     
-    func fetchMovieInformation(movieCode: String, complitionHandler: @escaping (Movie?) -> Void) {
-        let urlString = "\(MovieURL.detailURL)key=\(Key.key)&movieCd=\(movieCode)"
-        
-        executeRequest(with: urlString) { data in
-            guard let data = data else { return }
-            let safeData = DataDecoder.parseJson(type: Movie.self, data: data)
-            complitionHandler(safeData)
-        }
-    }
-    
-    private func fetchTodayDate() -> String {
-        let currentDate = Date()
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = DateFormat.yyyyMMdd
-        
-        guard let yesterday = Calendar.current.date(byAdding: .day, value: -1, to: currentDate) else { return .emptyString }
-        
-        let targetDate = dateFormatter.string(from: yesterday)
-        
-        return targetDate
-    }
-    
-    private func executeRequest(with urlString: String, complitionHandler: @escaping (Data?) -> Void) {
+    private func executeRequest<T: Decodable>(with urlString: String, type: T.Type, complitionHandler: @escaping (T?) -> Void) {
         guard let url = URL(string: urlString) else { return }
         
         let urlSession = URLSession(configuration: .default)
@@ -59,12 +30,14 @@ struct NetworkManager {
                 return
             }
             
-            complitionHandler(data)
+            
+            let safeData = parseJson(type: T.self, data: data)
+            complitionHandler(safeData)
         }
         task.resume()
     }
     
-    func parseJson<T: Decodable>(type: T.Type, data: Data) -> T? {
+    private func parseJson<T: Decodable>(type: T.Type, data: Data) -> T? {
         let decoder = JSONDecoder()
         do {
             let receivedData = try decoder.decode(type, from: data)
